+++ magik-typed-checks/src/main/java/nl/ramsolutions/sw/magik/typedchecks/checks/TypeDocTypeExistsTypeCheck.java
@@ -45,7 +45,7 @@ public class TypeDocTypeExistsTypeCheck extends MagikTypedCheck {
             .forEach(entry -> {
                 final AstNode typeNode = entry.getKey();
                 final TypeString typeString = entry.getValue();
-                final String message = String.format(MESSAGE, typeString);
+                final String message = String.format(MESSAGE, typeString.getFullString());
                 this.addIssue(typeNode, message);
             });
     }